move & rename the SASS compiler file
